Scanner mask (#92)
* Add scanner mask (use infinity value in LOR as indicator)

* Add chronological interleaved binIterator

# Conflicts:
#	yrt-pet/include/yrt-pet/datastruct/projection/BinIterator.hpp
#	yrt-pet/src/datastruct/projection/BinIterator.cpp
#	yrt-pet/unit_tests/utils/test_BinIterator.cpp

* Add first version of constrained bin iterator

* Fix scanner mask constraint, return from queue by reference

* Start histogram subset constraint

* Use base BinIterator (subsets), implement information collector

# Conflicts:
#	yrt-pet/include/utils/Tools.hpp
#	yrt-pet/src/utils/Tools.cpp

* Fix variable collection

* Update mask reading code

* Cache bin count

* Test bin iterator loop

# Conflicts:
#	yrt-pet/src/recon/OSEMUpdater_CPU.cpp

* Compilation fixes

* Add bin to projection properties, pass to corrector

* Reduce number of consumer threads

* Use std::thread and dynamic assign producers/consumers

* Convert string keys to enums

* Pass constraint information buffer to thread function by parameter

* Simple static scheduling, producers consume

* Compilation fixes

# Conflicts:
#	yrt-pet/src/recon/OSEMUpdater_CPU.cpp

* Remove queue

* Fixes after rebase

* Test projection properties type

* Create type for detector orientations, create projection property

* Start moving towards new projection property object

* Use flexible structure class for constraints

* Start change of ProjectionProperties type

* Propagate ProjectionProperties changes

* Update OSEM_CPU object

* Start move of GPU operators to new ProjectionProperties format

* Update method signatures

* Replace batch creation in ProjectionDataDevice object

# Conflicts:
#	yrt-pet/src/datastruct/projection/ProjectionDataDevice.cu

* Update calls to ProjectionDataDevice constructors

* Update OperatorProjectorDevice

* Add BinIteratorConstraints and buffers to projector

* Move Constraint class, start update of GPU code

* Update GPU projectors

* Fix executables and tests

* Fix type of ProjectionProperties buffer

* Formatting

* Delay BinIteratorConstrained setup, add numThreads field

* Add print function, fix warnings, default offset to -1

* Add LOR to required property types for GPU projectors

* Add atomic sum in backprojection (C++20), fix bug in applyA/applyAH

* Separate properties and constraint indices, fixes unit tests

* Use member variable to get number of threads

* Rename variable, fix test formatting

* Rename enum, add python bindings

* Fix OSEM updater code (pass thread ID to projector)

* Fix formatting

* Fix rebase

* Fix bug in OSEM setup: add TOF options

* Fix python interface for GPU projectors

* Rename enums, update parallelization, remove float4

* Fix bug in GPU multi-ray Siddon kernel

* Change index type to size_t

* Rename to BinFilter, adjust parameter type and name

* Account for corrections in memory calculation

* Allow large batch size (size_t), fix OSEM_GPU memory calculations

* Precalculate property collection flags

* Fix bin index type

* Fix size check when loading mask

* Add BinFilter test, fix periodicDiff

* Test constraints

# Conflicts:
#	yrt-pet/src/datastruct/scanner/DetRegular.cpp

* Add scanner_mask test, test different position, fix bugs

* Fix invalid scanner geometry

* Remove dependency on std::format

* Adjust tolerance after regular scanner update

* Remove include

* Remove unused object

* Document detMask scanner option

* Remove temporary code

---------

+++ yrt-pet/include/yrt-pet/datastruct/projection/ProjectionSpaceKernels.cuh
@@ -10,18 +10,18 @@ namespace yrt
 // kernels definitions
 __global__ void divideMeasurements_kernel(const float* d_dataIn,
                                           float* d_dataOut,
-                                          int maxNumberOfEvents);
+                                          size_t maxNumberOfEvents);
 __global__ void addProjValues_kernel(const float* d_dataIn, float* d_dataOut,
-                                     int maxNumberOfEvents);
+                                     size_t maxNumberOfEvents);
 __global__ void invertProjValues_kernel(const float* d_dataIn, float* d_dataOut,
-                                        int maxNumberOfEvents);
+                                        size_t maxNumberOfEvents);
 __global__ void convertToACFs_kernel(const float* d_dataIn, float* d_dataOut,
-                                     float unitFactor, int maxNumberOfEvents);
+                                     float unitFactor, size_t maxNumberOfEvents);
 __global__ void multiplyProjValues_kernel(const float* d_dataIn,
                                           float* d_dataOut,
-                                          int maxNumberOfEvents);
+                                          size_t maxNumberOfEvents);
 __global__ void multiplyProjValues_kernel(float scalar, float* d_dataOut,
-                                          int maxNumberOfEvents);
+                                          size_t maxNumberOfEvents);
 __global__ void clearProjections_kernel(float* d_dataIn, float value,
-                                        int maxNumberOfEvents);
+                                        size_t maxNumberOfEvents);
 }  // namespace yrt
